feat(http): preserve stream shutdown finalization
## What

- Resolve streaming codecs by full media type and reject unsupported media representations.
- Guarantee that Accept negotiation yields a response encoder, while bidirectional streams finalize both codecs before aborting during drain.

## Why

- Keep streaming protocol negotiation explicit and preserve codec cleanup when shutdown races with a committed handler failure.

## Testing

- `make package=net/http/content/stream specs` - passed
- `make package=net/http/client specs` - passed
- `make package=net/http/content/stream lint` - passed
- `make sec` - passed; govulncheck found no affected code and Trivy reported no vulnerabilities or secrets.
- `git diff --check` - passed
- Validation is scoped to the changed streaming package and its direct HTTP client consumer; CI covers the full suite.

AI-assisted-by: [Codex](https://openai.com/codex/)
