+++ net/http/content/stream/stream.go
@@ -158,6 +158,22 @@ func (s *Stream[Res]) close() error {
 	return s.encoder.Close()
 }
 
+// requestEncoder keeps bidirectional stream finalization on the response stream while closing both codecs.
+type requestEncoder struct {
+	stream.Encoder
+	decoder stream.Decoder
+}
+
+func (e *requestEncoder) Close() error {
+	if err := e.Encoder.Close(); err != nil {
+		_ = e.decoder.Close()
+
+		return err
+	}
+
+	return e.decoder.Close()
+}
+
 // RequestStream is a bidirectional streaming handle: both the request and the response stream.
 //
 // RequestStream is not safe for arbitrary concurrent use. The supported pattern, matching gRPC's bidi
@@ -265,14 +281,4 @@ func (s *RequestStream[Req, Res]) decode() (*Req, error) {
 // request stream, letting a handler's Recv loop end cleanly without importing errors/io itself.
 func (s *RequestStream[Req, Res]) IsFinished(err error) bool {
 	return errors.Is(err, io.EOF)
-}
-
-func (s *RequestStream[Req, Res]) close() error {
-	err := s.Stream.close()
-
-	if decErr := s.decoder.Close(); err == nil {
-		err = decErr
-	}
-
-	return err
 }
